MCF-06 Move all widgets from shared folder
